backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.7.4
anthropic==0.18.1
httpx==0.26.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
slowapi==0.1.9